feat: Add custom theme and update Storybook packages
This commit introduces a custom themed configuration and updates the Storybook packages in the "pkgs-storybook" app. A new theme file is created and applied through Storybook's manager.js file. Storybook's version is updated along with its dependencies, reflecting in package.json and pnpm-lock.yaml files. Also, the ".storybook/main.js" file is updated to include a new addon, and unnecessary files are removed as well.